cleanup GlobalUsings
