+++ src/Judge/NTS.Judge.Blazor/GlobalUsings.cs
@@ -1,5 +1,4 @@
 ﻿global using Microsoft.AspNetCore.Components;
 global using static Not.Localization.NStrings;
-global using static NTS.Application.DataTransferObjects.Models.CoreModel;
 global using NTS.Domain.Enums;
 global using static NTS.Localization.NtsStrings;
