+++ src/Judge/NTS.Judge/GlobalUsings.cs
@@ -1,4 +1,2 @@
 ﻿global using static Not.Localization.NStrings;
-global using static NTS.Application.DataTransferObjects.Models.CommonModel;
-global using static NTS.Application.DataTransferObjects.Models.CoreModel;
 global using static NTS.Localization.NtsStrings;
