+++ src/Judge/NTS.Judge/Nexus/Behinds/InquiryBehind.cs
@@ -1,5 +1,6 @@
 ﻿using Not.Application.Behinds.Adapters;
 using Not.Async;
+using NTS.Application.Models;
 using NTS.Domain.Core.Aggregates;
 using NTS.Judge.Blazor.Nexus;
 using NTS.Judge.Nexus.Repositories;
