+++ src/NTS.Application/Models/Document.cs
@@ -1,0 +1,16 @@
+﻿using Not.Domain;
+using Not.Storage.Tenants;
+using Not.Structures;
+
+namespace NTS.Application.Models;
+
+public abstract class Document : IDocument
+{
+    public int Id { get; init; }
+    public string TenantId { get; set; } = "nts";
+}
+
+// TODO: remove IAggregateRoot once only models are used in repositories and use IDocument in IRepository instead
+public interface IDocument : IIdentifiable, ITenantAware, IAggregateRoot
+{
+}
