+++ src/NTS.Application/Models/SetupModel.cs
@@ -1,0 +1,247 @@
+﻿using Not.DomainUtils;
+using NTS.Domain.Enums;
+using NTS.Domain.Setup.Aggregates;
+
+namespace NTS.Application.Models;
+
+public class SetupAthleteModel : AthleteModel, IDocument
+{
+    // TODO: if decide to use this approach integrate AutoMapper with specific mappings to solve duplicating mapping logic
+    public static SetupAthleteModel MapFrom(Athlete athlete)
+    {
+        return new SetupAthleteModel
+        {
+            Id = athlete.Id,
+            FeiId = athlete.FeiId,
+            Names = athlete.Names,
+            Country = CountryModel.MapFrom(athlete.Country),
+            Club = athlete.Club == null ? null : ClubModel.MapFrom(athlete.Club),
+        };
+    }
+
+    public string TenantId { get; init; } = StorageConstants.DEFAULT_TENANT;
+
+    public new Athlete MapToDomain()
+    {
+        var club = new Club(Club?.MapToDomain().Name);
+        return new Athlete(Id, Names, FeiId, Country?.MapToDomain(), club);
+    }
+}
+
+public class SetupCombinationModel
+{
+    public static SetupCombinationModel MapFrom(Combination combination)
+    {
+        return new SetupCombinationModel
+        {
+            Id = combination.Id,
+            Number = combination.Number,
+            Athlete = SetupAthleteModel.MapFrom(combination.Athlete),
+            Horse = HorseModel.MapFrom(combination.Horse),
+        };
+    }
+
+    public int Id { get; init; }
+    public int Number { get; init; }
+    public SetupAthleteModel Athlete { get; init; } = default!;
+    public HorseModel Horse { get; init; } = default!;
+
+    public Combination MapToDomain()
+    {
+        var athlete = Athlete.MapToDomain();
+        //intermediate horse will be removed after merge with codex/design-rpc-methods-for-nts.witness
+        var intermediateHorse = Horse.MaptoDomain();
+        var horse = new Horse(intermediateHorse.Id,intermediateHorse.Name,intermediateHorse.FeiId);
+        return new Combination(
+            DomainHelper.EnsureId(Id),
+            Number,
+            athlete,
+            horse,
+            null // Tag prop is not used currently
+        );
+    }
+}
+
+public class SetupParticipationModel
+{
+    public static SetupParticipationModel MapFrom(Participation participation)
+    {
+        return new SetupParticipationModel
+        {
+            Id = participation.Id,
+            IsNotRanked = participation.IsNotRanked,
+            Category = participation.Category,
+            Combination = SetupCombinationModel.MapFrom(participation.Combination),
+            StartTimeOverride = participation.StartTimeOverride,
+            MaxSpeedOverride = participation.MaxSpeedOverride,
+            MinSpeedOverride = participation.MinSpeedOverride,
+            MinAverageSpeed = participation.MinAverageSpeed,
+            MaxAverageSpeed = participation.MaxAverageSpeed,
+        };
+    }
+
+    public int Id { get; init; }
+    public ParticipationCategory Category { get; init; } = default!;
+    public SetupCombinationModel Combination { get; init; } = default!;
+    public bool IsNotRanked { get; init; }
+    public DateTimeOffset? StartTimeOverride { get; init; }
+    public double? MaxSpeedOverride { get; init; }
+    public double? MinSpeedOverride { get; init; }
+    public double? MinAverageSpeed { get; init; }
+    public double? MaxAverageSpeed { get; init; }
+
+    public Participation MapToDomain()
+    {
+        var combination = Combination.MapToDomain();
+        return new Participation(
+            DomainHelper.EnsureId(Id),
+            IsNotRanked,
+            combination,
+            Category,
+            StartTimeOverride,
+            MaxSpeedOverride,
+            MinSpeedOverride,
+            MinAverageSpeed,
+            MaxAverageSpeed
+        );
+    }
+}
+
+public class LoopModel
+{
+    public static LoopModel MapFrom(Loop loop)
+    {
+        return new() { Id = loop.Id, Distance = loop.Distance };
+    }
+
+    public int Id { get; init; }
+    public double Distance { get; init; }
+
+    public Loop MapToDomain()
+    {
+        return Loop.Update(DomainHelper.EnsureId(Id), Distance);
+    }
+}
+
+public class SetupPhaseModel
+{
+    public static SetupPhaseModel Create(Phase phase)
+    {
+        return new SetupPhaseModel
+        {
+            Loop = LoopModel.MapFrom(phase.Loop!),
+            Recovery = phase.Recovery,
+            Rest = phase.Rest,
+        };
+    }
+
+    public int Recovery { get; init; }
+    public LoopModel Loop { get; init; } = default!;
+    public int? Rest { get; init; }
+
+    public Phase MapToDomain()
+    {
+        return new Phase(Loop.MapToDomain(), Recovery, Rest);
+    }
+}
+
+public class SetupCompetitionModel
+{
+    public static SetupCompetitionModel MapFrom(Competition competition)
+    {
+        return new SetupCompetitionModel
+        {
+            Id = competition.Id,
+            Name = competition.Name,
+            Type = competition.Type,
+            Ruleset = competition.Ruleset,
+            Start = competition.Start,
+            CompulsoryThreshold = competition.CompulsoryThresholdSpan,
+            FeiId = competition.FeiId,
+            FeiRule = competition.FeiRule,
+            FeiScheduleNumber = competition.FeiScheduleNumber,
+            Phases = competition.Phases.Select(SetupPhaseModel.Create).ToArray(),
+            Participations = competition.Participations.Select(SetupParticipationModel.MapFrom).ToArray(),
+        };
+    }
+
+    public int Id { get; init; }
+    public string Name { get; init; } = default!;
+    public CompetitionType Type { get; init; }
+    public CompetitionRuleset Ruleset { get; init; }
+    public DateTimeOffset? Start { get; init; }
+    public TimeSpan? CompulsoryThreshold { get; init; }
+    public string? FeiId { get; init; }
+    public string? FeiRule { get; init; }
+    public string? FeiScheduleNumber { get; init; }
+    public SetupPhaseModel[] Phases { get; init; } = default!;
+    public SetupParticipationModel[] Participations { get; init; } = default!;
+
+    public Competition MapToDomain()
+    {
+        return new Competition(
+            DomainHelper.EnsureId(Id),
+            Name,
+            Type,
+            Ruleset,
+            Start,
+            CompulsoryThreshold,
+            FeiId,
+            FeiRule,
+            FeiScheduleNumber,
+            Phases.Select(x => x.MapToDomain()),
+            Participations.Select(x => x.MapToDomain())
+        );
+    }
+}
+
+public class UpcomingEventModel : IDocument
+{
+    public static UpcomingEventModel MapFrom(UpcomingEvent @event)
+    {
+        return new UpcomingEventModel
+        {
+            Id = @event.Id,
+            Name = @event.Name,
+            Place = @event.Place,
+            Country = CountryModel.MapFrom(@event.Country),
+            ShowFeiId = @event.ShowFeiId,
+            FeiId = @event.FeiId,
+            FeiEventCode = @event.FeiEventCode,
+            Competitions = @event.Competitions.Select(SetupCompetitionModel.MapFrom).ToArray(),
+            Officials = @event.Officials.Select(OfficialModel.MapFrom).ToArray(),
+            Loops = @event.Loops.Select(LoopModel.MapFrom).ToArray(),
+            Combinations = @event.Combinations.Select(SetupCombinationModel.MapFrom).ToArray(),
+        };
+    }
+
+    public int Id { get; init; } = default!;
+    public string TenantId { get; init; } = StorageConstants.DEFAULT_TENANT;
+    public string Place { get; init; } = default!;
+    public CountryModel Country { get; init; } = default!;
+    public string? ShowFeiId { get; init; }
+    public string? FeiId { get; init; }
+    public string? FeiEventCode { get; init; }
+    public SetupCompetitionModel[] Competitions { get; init; } = default!;
+    public OfficialModel[] Officials { get; init; } = default!;
+    public LoopModel[] Loops { get; init; } = default!;
+    public SetupCombinationModel[] Combinations { get; init; } = default!;
+    public string Name { get; init; } = default!;
+
+    public UpcomingEvent MapToDomain()
+    {
+        return new UpcomingEvent(
+            Id,
+            Name,
+            Place,
+            Country.MapToDomain(),
+            ShowFeiId,
+            FeiId,
+            FeiEventCode,
+            Competitions.Select(x => x.MapToDomain()),
+            Officials.Select(x => x.MapToSetupDomain()),
+            Loops.Select(x => x.MapToDomain()),
+            Combinations.Select(x => x.MapToDomain())
+        );
+    }
+}
