+++ src/NTS.Application/StorageConstants.cs
@@ -1,0 +1,5 @@
+﻿namespace NTS.Application;
+internal static class StorageConstants
+{
+    public const string DEFAULT_TENANT = "nts";
+}
