+++ src/NTS.Domain.Core/Aggregates/Participations/Athlete.cs
@@ -6,7 +6,7 @@ public class Athlete : IAthlete
 {
     [Newtonsoft.Json.JsonConstructor]
     [System.Text.Json.Serialization.JsonConstructor]
-    public Athlete(int id, Person names, Country country, Club? club, string feiId)
+    public Athlete(int id, Person names, Country country, Club? club, string? feiId)
     {
         Id = id;
         Names = names;
