+++ src/NTS.Warp/ACL/Factories/CompetitionFactory.cs
@@ -1,4 +1,6 @@
-﻿using Not.Extensions;
+﻿using Not.DomainUtils;
+using Not.Extensions;
+using NTS.Application.Models;
 using NTS.Domain.Enums;
 using NTS.Warp.ACL.Entities.Competitions;
 using NTS.Warp.ACL.Models;
@@ -7,13 +9,13 @@ namespace NTS.Warp.ACL.Factories;
 
 public class CompetitionFactory
 {
-    public static EmsCompetition Create(ParticipationModel participation)
+    public static EmsCompetition Create(CoreParticipationModel participation)
     {
         if (participation.Competition == null)
         {
             var emptyState = new EmsCompetitionState
             {
-                Id = DomainModelHelper.GenerateId(),
+                Id = DomainHelper.GenerateId(),
                 Name = "Unknown Competition",
                 Type = EmsCompetitionType.National,
             };
@@ -22,7 +24,7 @@ public class CompetitionFactory
         var laps = LapFactory.Create(participation.Phases);
         var state = new EmsCompetitionState
         {
-            Id = participation.Competition.Id,
+            Id = DomainHelper.GenerateId(),
             Name = participation.Competition.Name,
             Type = MapEmsCompetitionType(participation.Competition.Ruleset),
         };
