+++ src/NTS.Warp/ACL/Factories/ParticipantEntryFactory.cs
@@ -1,10 +1,11 @@
-﻿using NTS.Warp.ACL.Entities;
+﻿using NTS.Application.Models;
+using NTS.Warp.ACL.Entities;
 
 namespace NTS.Warp.ACL.Factories;
 
 public class ParticipantEntryFactory
 {
-    public static EmsParticipantEntry Create(ParticipationModel participation)
+    public static EmsParticipantEntry Create(CoreParticipationModel participation)
     {
         var emsParticipation = ParticipationFactory.CreateEms(participation);
         return new EmsParticipantEntry(emsParticipation);
