+++ src/NTS.Warp/ACL/Factories/StartlistFactory.cs
@@ -1,10 +1,11 @@
+using NTS.Application.Models;
 using NTS.Warp.ACL.Entities;
 
 namespace NTS.Warp.ACL.Factories;
 
 public static class StartlistFactory
 {
-    public static Dictionary<int, EmsStartlist> Create(IEnumerable<ParticipationModel> participations)
+    public static Dictionary<int, EmsStartlist> Create(IEnumerable<CoreParticipationModel> participations)
     {
         var emsParticipations = participations.Select(ParticipationFactory.CreateEms);
         var startlists = new Dictionary<int, EmsStartlist>();
