+++ src/NTS.Warp/Features/Judge/Procedures/IParticipationClientProcedures.cs
@@ -1,3 +1,4 @@
+using NTS.Application.Models;
 using NTS.Domain.Objects;
 
 namespace NTS.Warp.Features.Judge.Procedures;
@@ -5,5 +6,5 @@ namespace NTS.Warp.Features.Judge.Procedures;
 public interface IParticipationClientProcedures
 {
     Task ProcessSnapshots(IEnumerable<Snapshot> snapshots);
-    Task<IEnumerable<ParticipationModel>> GetActiveParticipations();
+    Task<IEnumerable<CoreParticipationModel>> GetActiveParticipations();
 }
