--- src/NTS.Warp/GlobalUsings.cs
@@ -1,3 +1,0 @@
-﻿global using static NTS.Application.DataTransferObjects.Models.CommonModel;
-global using static NTS.Application.DataTransferObjects.Models.CoreModel;
-global using static NTS.Application.DataTransferObjects.Models.SetupModel;
