+++ src/Nexus/NTS.Nexus.HTTP/Functions/Countries/CountryFunctions.cs
@@ -28,7 +28,7 @@ public class CountryFunctions : FunctionBase<CountryFunctions>
 
         var requestBody = await new StreamReader(request.Body).ReadToEndAsync();
         var country = requestBody.FromJson<Country>();
-        var document = CountryModel.Create(country);
+        var document = CountryModel.MapFrom(country);
         await _countries.Create(document);
 
         return new OkObjectResult($"Inserted {country}");
@@ -43,7 +43,7 @@ public class CountryFunctions : FunctionBase<CountryFunctions>
 
         var requestBody = await new StreamReader(request.Body).ReadToEndAsync();
         var country = requestBody.FromJson<Country>();
-        var document = CountryModel.Create(country);
+        var document = CountryModel.MapFrom(country);
         await _countries.Update(document);
 
         return new OkObjectResult($"Updated {country}");
@@ -56,7 +56,7 @@ public class CountryFunctions : FunctionBase<CountryFunctions>
     {
         LogInformation(request);
 
-        var countries = await _countries.ReadAll().Select(x => x.ToDomain());
+        var countries = await _countries.ReadAll().Select(x => x.MapToDomain());
         return new OkObjectResult(countries);
     }
 }
