+++ src/Nexus/NTS.Nexus.HTTP/GlobalUsings.cs
@@ -1,5 +1,2 @@
 ﻿global using static Not.Localization.NStrings;
-global using static NTS.Application.DataTransferObjects.Models.CommonModel;
-global using static NTS.Application.DataTransferObjects.Models.CoreModel;
-global using static NTS.Application.DataTransferObjects.Models.SetupModel;
 global using static NTS.Localization.NtsStrings;
